+++ src/core/som.h
@@ -17,7 +17,7 @@ using namespace std;
  */
 class SOM {
 public:
-    vector<vector<float> > input_space;
+    vector<vector<double> > input_space;
     Grid* grid;
     int current_iteration;
     int selected_input_vector_index;
@@ -36,13 +36,13 @@ public:
      * @param is_random Random selection of the next input vector
      * @return The next input vector
      */
-    vector<float> GetNextInputVector(bool is_random = false);
+    vector<double> GetNextInputVector(bool is_random = false);
 
     /**
      * Start training
      * @param input_space Input space for the train
      */
-    void Train(vector<vector<float> > &input_space);
+    void Train(vector<vector<double> > &input_space);
 };
 
 #endif //SOM_TRAINER_H
