+++ test/grid_test.cpp
@@ -28,12 +28,12 @@ TEST(FindBMU, Manual){
     Grid* grid = new Rectangular(2, 2, 3, false);
     grid->InitializeNodes();
 
-    grid->nodes[0].weight_vector = vector<float>{10, 20, 30};
-    grid->nodes[1].weight_vector = vector<float>{-10, 2, 7};
-    grid->nodes[2].weight_vector = vector<float>{1, 2, 5};
-    grid->nodes[3].weight_vector = vector<float>{-5, -4, -3};
+    grid->nodes[0].weight_vector = vector<double>{10, 20, 30};
+    grid->nodes[1].weight_vector = vector<double>{-10, 2, 7};
+    grid->nodes[2].weight_vector = vector<double>{1, 2, 5};
+    grid->nodes[3].weight_vector = vector<double>{-5, -4, -3};
 
-    vector<float>vec = {3, 4, 5};
+    vector<double>vec = {3, 4, 5};
     int bmu = grid->FindBMU(vec);
 
     ASSERT_EQ(bmu, 2);
